Tests for CustomerRepository
