+++ 2020/BillSplitter/BillSplitterTests/BillSplitterTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -7,7 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="CustomerDbAccessorTests.cs" />
     <Compile Remove="OrdersDbAccessorTests.cs" />
     <Compile Remove="PositionsDbAccessorTests.cs" />
     <Compile Remove="UsersDbAccessorTests.cs" />
